TDL - 18007: Dict based to class-based refactoring (#21)
* refactoring dict based to class based

* removed unnecessary logger and resolved parent child bug

* Added comments for functions

* Resolved endpoint conflict issue

* Removed pagination for devices and platform stream

* Removed improper test cases

* Resolved pylint errors.
